HBAPP-33 Move model load menu to Playground. Add a playground tab to main menu. Adjust wording for performance settings menu. Capitalize tab labels.
