+++ components/chat-container.tsx
@@ -8,19 +8,7 @@ import { useSettings } from '@/components/features/settings/hooks'
 import { ModelID } from '@/components/features/settings/types'
 import { Button } from '@/components/ui/button'
 import { toast } from 'react-hot-toast'
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue
-} from '@/components/ui/select'
-import { ResponseCharmMenu } from '@/components/features/prompt/dialog-response-charm'
 import { ApplicationModesMenu } from '@/components/features/menus/menu-application-modes'
-import { MixerHorizontalIcon, LightningBoltIcon } from '@radix-ui/react-icons'
-import { I_LLM_Options } from '@/lib/hooks/types'
 
 interface IProps {
   id?: string
@@ -38,11 +26,8 @@ export const ChatContainer = ({ id, initialMessages }: IProps) => {
   const { provider: selectedProvider, model: selectedModel } = useSettings()
   const { connect: connectToHomebrew, getServices } = useHomebrew()
   const [currentTextModel, setCurrentTextModel] = useState(null)
-  const [selectedModelId, setSelectedModelId] = useState<string | undefined>(undefined)
   const [installedList, setInstalledList] = useState<T_InstalledTextModel[]>([])
   const [modelConfigs, setModelConfigs] = useState<{ [key: string]: T_ModelConfig }>()
-  const [openResponseCharmDialog, setOpenResponseCharmDialog] = useState(false)
-  const [responseSettings, setResponseSettings] = useState(null)
 
   const connect = useCallback(async () => {
     setIsConnecting(true)
@@ -77,63 +62,9 @@ export const ChatContainer = ({ id, initialMessages }: IProps) => {
     return false
   }, [connectToHomebrew, getServices, selectedProvider])
 
-  const connectTextServiceAction = useCallback(async () => {
-    const action = async () => {
-      try {
-        // First check if a model is already loaded, if so skip...
-        const modelResponse = await services?.textInference.model()
-        if (modelResponse?.success) {
-          toast.success(`Success: ${modelResponse?.message}`)
-          return true
-        }
-        // Pass any settings data we find, We could instead pass init args from a user input, using saved settings for now.
-        const settingsResponse = await services?.storage.getSettings()
-        if (!settingsResponse?.success) {
-          toast.error(`${settingsResponse?.message}`)
-        }
-        // Remove "preset" from init payload
-        const initOptions = { ...settingsResponse?.data?.init }
-        if (initOptions?.preset) delete initOptions['preset']
-        // Set "call" payload
-        const callOptions = { ...settingsResponse?.data?.call }
-        // Tell backend to load the model into memory using these args
-        const installPath = installedList?.find(i => i.id === selectedModelId)?.savePath
-        const mode = 'completion' // @TODO Set chat "mode" in payload from UI
-        const payload = { modelPath: installPath, modelId: selectedModelId, mode, init: initOptions, call: callOptions }
-        const response = await services?.textInference.load({ body: payload })
-
-        if (response?.success) {
-          toast.success('Connected successfully to Ai')
-          setCurrentTextModel(response?.data)
-          return true
-        }
-
-        toast.error(response?.message || 'Failed to connect to Ai.')
-        return false
-      } catch (error) {
-        toast.error(`${error}`)
-        return false
-      }
-    }
-
-    setIsConnecting(true)
-    const result = await action()
-    setIsConnecting(false)
-    return result
-  }, [installedList, selectedModelId, services?.storage, services?.textInference])
-
   const isLocalSelected = selectedProvider === ModelID.Local
   const isCloudSelected = selectedProvider !== ModelID.Local && selectedModel !== 'no model selected'
 
-  const installedModels = installedList?.map(item => {
-    const cfg = modelConfigs?.[item.id]
-    const name = cfg?.name
-    return (<SelectItem key={item.id} value={item.id}>{name}</SelectItem>)
-  })
-
-  const fetchSettings = useCallback(async () => services?.storage.getSettings(), [services?.storage])
-
-  const saveSettings = async (options: I_LLM_Options) => services?.storage.saveSettings({ body: options })
 
   // HomeBrewAi connection menu
   if (!isConnected)
@@ -153,70 +84,11 @@ export const ChatContainer = ({ id, initialMessages }: IProps) => {
   // Inference connection menu
   if (!hasTextServiceConnected)
     return (
-      <>
-        {/* Dialog Menu for Response settings */}
-        <ResponseCharmMenu
-          dialogOpen={openResponseCharmDialog}
-          setDialogOpen={setOpenResponseCharmDialog}
-          onSubmit={(charm, settings) => {
-            saveSettings(settings)
-            toast.success('Model settings saved!')
-          }}
-          settings={responseSettings}
-          modelConfig={modelConfigs?.[selectedModelId || '']}
-        />
-        {/* Model Selection Menu */}
-        <div className="flex w-full flex-col gap-16 overflow-hidden p-4 md:w-[70%]">
-          <div className="mt-4 px-4 text-center">Connected to HomebrewAi server</div>
-          <div className="flex flex-col items-stretch justify-items-stretch gap-4">
-            {/* Start */}
-            {selectedModelId && <Button
-              className="h-fit min-w-fit flex-1 bg-blue-600 px-8 text-center text-white hover:bg-blue-800"
-              onClick={async () => {
-                setIsConnecting(true)
-                const isConnected = await connectTextServiceAction()
-                isConnected && setHasTextServiceConnected(true)
-                setIsConnecting(false)
-              }}
-              disabled={isConnecting}
-            >
-              <LightningBoltIcon className="mr-1" />Start
-            </Button>}
-            <div className="flex flex-row gap-2">
-              {/* Select a prev installed model to load */}
-              <div className="w-full">
-                <Select
-                  defaultValue={undefined}
-                  value={selectedModelId}
-                  onValueChange={setSelectedModelId}
-                >
-                  <SelectTrigger className="w-full flex-1">
-                    <SelectValue placeholder="Select Ai Model"></SelectValue>
-                  </SelectTrigger>
-                  <SelectGroup>
-                    <SelectContent className="p-1">
-                      <SelectLabel className="select-none uppercase text-indigo-500">Installed</SelectLabel>
-                      {installedModels}
-                    </SelectContent>
-                  </SelectGroup>
-                </Select>
-              </div>
-              {/* Model Settings Button */}
-              {selectedModelId && <Button
-                className="m-auto h-fit bg-accent-foreground hover:bg-accent"
-                variant="outline"
-                onClick={
-                  async () => {
-                    await fetchSettings().then(res => setResponseSettings(res?.data?.init))
-                    setOpenResponseCharmDialog(true)
-                  }}>
-                <MixerHorizontalIcon className="mr-1" />Settings
-              </Button>}
-            </div>
-          </div>
-          <ApplicationModesMenu onSubmit={() => {/* logic to go to a route */ }} />
-        </div>
-      </>
+      // Model Selection Menu
+      < div className="flex w-full flex-col overflow-hidden p-4 md:w-[70%]" >
+        {/* <h1 className="mt-4 px-4 text-center text-xl">Connected to HomebrewAi server</h1> */}
+        <ApplicationModesMenu setHasTextServiceConnected={setHasTextServiceConnected} isConnecting={isConnecting} setIsConnecting={setIsConnecting} modelConfigs={modelConfigs} installedList={installedList} setCurrentTextModel={setCurrentTextModel} onSubmit={() => {/* logic to go to a route */ }} services={services} />
+      </div >
     )
 
   // Render "Connecting..." feedback
